refactor: split code into smaller files
