+++ internal/ui/messages.go
@@ -1,0 +1,66 @@
+package ui
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	tea "github.com/charmbracelet/bubbletea"
+
+	"storyblok-sync/internal/sb"
+)
+
+// ---------- Messages / Cmds ----------
+type validateMsg struct {
+	spaces []sb.Space
+	err    error
+}
+
+type scanMsg struct {
+	src []sb.Story
+	tgt []sb.Story
+	err error
+}
+
+func (m Model) validateTokenCmd() tea.Cmd {
+	return func() tea.Msg {
+		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+		defer cancel()
+		c := sb.New(m.cfg.Token)
+		spaces, err := c.ListSpaces(ctx)
+		if err != nil {
+			return validateMsg{err: err}
+		}
+		return validateMsg{spaces: spaces, err: nil}
+	}
+}
+
+func (m Model) scanStoriesCmd() tea.Cmd {
+	srcID, tgtID := 0, 0
+	if m.sourceSpace != nil {
+		srcID = m.sourceSpace.ID
+	}
+	if m.targetSpace != nil {
+		tgtID = m.targetSpace.ID
+	}
+	token := m.cfg.Token
+
+	return func() tea.Msg {
+		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+		defer cancel()
+		c := sb.New(token)
+
+		// Parallel wäre nice-to-have, hier sequentiell für Klarheit
+		src, err := c.ListStories(ctx, sb.ListStoriesOpts{SpaceID: srcID, PerPage: 50})
+		if err != nil {
+			return scanMsg{err: fmt.Errorf("source scan: %w", err)}
+		}
+		sortStories(src)
+		tgt, err := c.ListStories(ctx, sb.ListStoriesOpts{SpaceID: tgtID, PerPage: 50})
+		if err != nil {
+			return scanMsg{err: fmt.Errorf("target scan: %w", err)}
+		}
+		sortStories(tgt)
+		return scanMsg{src: src, tgt: tgt, err: nil}
+	}
+}
